release version

+++ Sort.h
@@ -19,6 +19,7 @@ public:
 	bool mergeSortTick();
 	bool heapSortTick();
 	bool quickSortTick();
+
 	void drawArray();
 	int getOperationCounter();
 	int getArraySize();
@@ -29,20 +30,19 @@ private:
 	int max_value; //max value of array elements
 	int x, y;	//drawArray bottom left coordinates
 	float x_scale, y_scale;	//size scale
-
 	int n;  //size of subarray
 	int i;  //current iteration
 	int operation_counter;  //number of swap operation done
 
 	//Merge sort fields
-	size_t BlockSizeIterator;
-	size_t BlockIterator;
-	size_t LeftBlockIterator;
-	size_t RightBlockIterator;
-	size_t MergeIterator;
-	size_t LeftBorder;
-	size_t MidBorder;
-	size_t RightBorder;
+	int BlockSizeIterator;
+	int BlockIterator;
+	int LeftBlockIterator;
+	int RightBlockIterator;
+	int MergeIterator;
+	int LeftBorder;
+	int MidBorder;
+	int RightBorder;
 	std::vector <int>SortedBlock;
 
 	//Heap sort fields
@@ -50,8 +50,8 @@ private:
 	int root;
 
 	//Quick sort fields
-	size_t l;
-	size_t h;
+	int l;
+	int h;
 	std::vector <int> stack;
 	int top;
 	int j;
